05-data-pipelines/airflow/requirements.txt
pyspark==3.3.0
tensorflow==2.9.1
reportlab==3.6.11
marshmallow-sqlalchemy==0.26.1
aiobotocore==2.4.0
botocore==1.27.59
boto==2.49.0
boto3==1.24.31
mypy-boto3-appflow==1.24.90
mypy-boto3-rds==1.24.0
mypy-boto3-redshift-data==1.24.11.post3
faker
apache-airflow==2.4.1
airflow-dbt==0.4.0
airflow-provider-great-expectations==0.1.4
apache-airflow-providers-amazon==6.0.0
apache-airflow-providers-common-sql==1.2.0
apache-airflow-providers-ftp==2.1.2
apache-airflow-providers-http==2.1.2
apache-airflow-providers-imap==2.2.3
psycopg2-binary==2.9.3
apache-airflow-providers-postgres==5.2.2
apache-airflow-providers-sqlite==2.1.3
s3fs==2022.8.2
snowflake-sqlalchemy==1.4.3
snowflake-connector-python==2.7.9
pretty-html-table
great-expectations==0.15.13
apache-airflow[postgres]
apache-airflow-providers-apache-spark